# openapi.yaml
openapi: 3.0.3
info:
  title: pglet API
  version: 0.1.0
  description: PostgreSQL database browser API
paths:
  /api/connect:
    post:
      operationId: connect
      summary: Connect to a PostgreSQL database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRequest'
      responses:
        '200':
          description: Connection info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionInfo'
        '400':
          description: Connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/switchdb:
    post:
      operationId: switchDatabase
      summary: Switch to a different database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchDBRequest'
      responses:
        '200':
          description: Connection info for new database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionInfo'
...
  /api/query:
    post:
      operationId: runQuery
      summary: Execute SQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResult'
  /api/explain:
    post:
      operationId: explainQuery
      summary: EXPLAIN a SQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Explain result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResult'
  /api/analyze:
    post:
      operationId: analyzeQuery
      summary: EXPLAIN ANALYZE a SQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Analyze result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResult'
  /api/query/cancel:
    post:
      operationId: cancelQuery
      summary: Cancel a running query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/export:
    post:
      operationId: exportQuery
      summary: Export query results as CSV or JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Exported data as CSV or JSON
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Query execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/saved-queries:
    get:
      operationId: listSavedQueries
      summary: List saved queries
      parameters:
        - name: database
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Saved query list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedQuery'
    post:
      operationId: createSavedQuery
      summary: Create a saved query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedQueryInput'
      responses:
        '201':
          description: Created saved query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedQuery'
  /api/saved-queries/{id}:
    get:
      operationId: getSavedQuery
      summary: Get a saved query by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saved query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedQuery'
    put:
      operationId: updateSavedQuery
      summary: Update a saved query
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedQueryInput'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      operationId: deleteSavedQuery
      summary: Delete a saved query
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/history:
    get:
      operationId: listHistory
      summary: List query history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: History entries with total count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
    delete:
      operationId: clearHistory
      summary: Clear all query history
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/ai/generate:
    post:
      operationId: aiGenerate
      summary: Generate SQL from natural language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiGenerateRequest'
      responses:
        '200':
          description: Generated SQL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiGenerateResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
    ConnectRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
    SwitchDBRequest:
      type: object
      required: [database]
      properties:
        database:
          type: string
    ConnectionInfo:
      type: object
      required: [host, port, user, database, version]
      properties:
        host:
          type: string
        port:
          type: integer
        user:
          type: string
        database:
          type: string
        version:
          type: string
        connected:
          type: boolean
...
    CellValue:
      type: string
      nullable: true
      x-go-type: '*string'
    QueryResult:
      type: object
      required: [columns, column_types, rows, row_count, duration_ms]
      properties:
        columns:
          type: array
          items:
            type: string
        column_types:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/CellValue'
        row_count:
          type: integer
        duration_ms:
          type: integer
          format: int64
        error:
          type: string
...
    QueryRequest:
      type: object
      required: [query, tab_id]
      properties:
        query:
          type: string
        tab_id:
          type: string
    CancelRequest:
      type: object
      required: [tab_id]
      properties:
        tab_id:
          type: string
    ExportRequest:
      type: object
      required: [query, format]
      properties:
        query:
          type: string
        format:
          type: string
          enum: [csv, json]
    SavedQuery:
      type: object
      required: [id, title, description, sql, database, tags, shared, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        sql:
          type: string
        database:
          type: string
        tags:
          type: string
        shared:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    SavedQueryInput:
      type: object
      required: [title, sql]
      properties:
        title:
          type: string
        description:
          type: string
          default: ''
        sql:
          type: string
        database:
          type: string
          default: ''
        tags:
          type: string
          default: ''
    HistoryEntry:
      type: object
      required: [id, sql, database, duration_ms, row_count, error, executed_at]
      properties:
        id:
          type: integer
        sql:
          type: string
        database:
          type: string
        duration_ms:
          type: integer
          format: int64
        row_count:
          type: integer
        error:
          type: string
        executed_at:
          type: string
    HistoryResponse:
      type: object
      required: [entries, total]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/HistoryEntry'
        total:
          type: integer
...
    AiGenerateRequest:
      type: object
      required: [prompt, database]
      properties:
        prompt:
          type: string
        database:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AiMessage'
    AiGenerateResponse:
      type: object
      required: [sql, explanation]
      properties:
        sql:
          type: string
        explanation:
          type: string
    AiMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string